+++ src/main/java/pt/andronikus/entities/base/LifeCycleTransitionsCost.java
@@ -1,196 +1,196 @@
 package pt.andronikus.entities.base;
 
 public class LifeCycleTransitionsCost {
-    private Float liveToSuspend;
-    private Float liveToStopped;
-    private Float stoppedToLive;
-    private Float stoppedToSuspend;
-    private Float stoppedToStandby;
-    private Float preActiveToStopped;
-    private Float standbyToSuspend;
-    private Float suspendToStandby;
-    private Float standbyToStopped;
-    private Float suspendToStopped;
-    private Float suspendToLive;
-    private Float standbyToLive;
-    private Float liveToStandby;
-    private Float testToStopped;
-    private Float preActiveToTest;
-    private Float testToLive;
-    private Float preActiveToLive;
-    private Float admActiveToCanceled;
-    private Float admActiveToSuspended;
-    private Float admSuspendedToActive;
-    private Float admSuspendedToCanceled;
+    private Double liveToSuspend;
+    private Double liveToStopped;
+    private Double stoppedToLive;
+    private Double stoppedToSuspend;
+    private Double stoppedToStandby;
+    private Double preActiveToStopped;
+    private Double standbyToSuspend;
+    private Double suspendToStandby;
+    private Double standbyToStopped;
+    private Double suspendToStopped;
+    private Double suspendToLive;
+    private Double standbyToLive;
+    private Double liveToStandby;
+    private Double testToStopped;
+    private Double preActiveToTest;
+    private Double testToLive;
+    private Double preActiveToLive;
+    private Double admActiveToCanceled;
+    private Double admActiveToSuspended;
+    private Double admSuspendedToActive;
+    private Double admSuspendedToCanceled;
 
     public LifeCycleTransitionsCost() {
     }
 
-    public Float getLiveToSuspend() {
+    public Double getLiveToSuspend() {
         return liveToSuspend;
     }
 
-    public void setLiveToSuspend(Float liveToSuspend) {
+    public void setLiveToSuspend(Double liveToSuspend) {
         this.liveToSuspend = liveToSuspend;
     }
 
-    public Float getLiveToStopped() {
+    public Double getLiveToStopped() {
         return liveToStopped;
     }
 
-    public void setLiveToStopped(Float liveToStopped) {
+    public void setLiveToStopped(Double liveToStopped) {
         this.liveToStopped = liveToStopped;
     }
 
-    public Float getStoppedToLive() {
+    public Double getStoppedToLive() {
         return stoppedToLive;
     }
 
-    public void setStoppedToLive(Float stoppedToLive) {
+    public void setStoppedToLive(Double stoppedToLive) {
         this.stoppedToLive = stoppedToLive;
     }
 
-    public Float getStoppedToSuspend() {
+    public Double getStoppedToSuspend() {
         return stoppedToSuspend;
     }
 
-    public void setStoppedToSuspend(Float stoppedToSuspend) {
+    public void setStoppedToSuspend(Double stoppedToSuspend) {
         this.stoppedToSuspend = stoppedToSuspend;
     }
 
-    public Float getStoppedToStandby() {
+    public Double getStoppedToStandby() {
         return stoppedToStandby;
     }
 
-    public void setStoppedToStandby(Float stoppedToStandby) {
+    public void setStoppedToStandby(Double stoppedToStandby) {
         this.stoppedToStandby = stoppedToStandby;
     }
 
-    public Float getPreActiveToStopped() {
+    public Double getPreActiveToStopped() {
         return preActiveToStopped;
     }
 
-    public void setPreActiveToStopped(Float preActiveToStopped) {
+    public void setPreActiveToStopped(Double preActiveToStopped) {
         this.preActiveToStopped = preActiveToStopped;
     }
 
-    public Float getStandbyToSuspend() {
+    public Double getStandbyToSuspend() {
         return standbyToSuspend;
     }
 
-    public void setStandbyToSuspend(Float standbyToSuspend) {
+    public void setStandbyToSuspend(Double standbyToSuspend) {
         this.standbyToSuspend = standbyToSuspend;
     }
 
-    public Float getSuspendToStandby() {
+    public Double getSuspendToStandby() {
         return suspendToStandby;
     }
 
-    public void setSuspendToStandby(Float suspendToStandby) {
+    public void setSuspendToStandby(Double suspendToStandby) {
         this.suspendToStandby = suspendToStandby;
     }
 
-    public Float getStandbyToStopped() {
+    public Double getStandbyToStopped() {
         return standbyToStopped;
     }
 
-    public void setStandbyToStopped(Float standbyToStopped) {
+    public void setStandbyToStopped(Double standbyToStopped) {
         this.standbyToStopped = standbyToStopped;
     }
 
-    public Float getSuspendToStopped() {
+    public Double getSuspendToStopped() {
         return suspendToStopped;
     }
 
-    public void setSuspendToStopped(Float suspendToStopped) {
+    public void setSuspendToStopped(Double suspendToStopped) {
         this.suspendToStopped = suspendToStopped;
     }
 
-    public Float getSuspendToLive() {
+    public Double getSuspendToLive() {
         return suspendToLive;
     }
 
-    public void setSuspendToLive(Float suspendToLive) {
+    public void setSuspendToLive(Double suspendToLive) {
         this.suspendToLive = suspendToLive;
     }
 
-    public Float getStandbyToLive() {
+    public Double getStandbyToLive() {
         return standbyToLive;
     }
 
-    public void setStandbyToLive(Float standbyToLive) {
+    public void setStandbyToLive(Double standbyToLive) {
         this.standbyToLive = standbyToLive;
     }
 
-    public Float getLiveToStandby() {
+    public Double getLiveToStandby() {
         return liveToStandby;
     }
 
-    public void setLiveToStandby(Float liveToStandby) {
+    public void setLiveToStandby(Double liveToStandby) {
         this.liveToStandby = liveToStandby;
     }
 
-    public Float getTestToStopped() {
+    public Double getTestToStopped() {
         return testToStopped;
     }
 
-    public void setTestToStopped(Float testToStopped) {
+    public void setTestToStopped(Double testToStopped) {
         this.testToStopped = testToStopped;
     }
 
-    public Float getPreActiveToTest() {
+    public Double getPreActiveToTest() {
         return preActiveToTest;
     }
 
-    public void setPreActiveToTest(Float preActiveToTest) {
+    public void setPreActiveToTest(Double preActiveToTest) {
         this.preActiveToTest = preActiveToTest;
     }
 
-    public Float getTestToLive() {
+    public Double getTestToLive() {
         return testToLive;
     }
 
-    public void setTestToLive(Float testToLive) {
+    public void setTestToLive(Double testToLive) {
         this.testToLive = testToLive;
     }
 
-    public Float getPreActiveToLive() {
+    public Double getPreActiveToLive() {
         return preActiveToLive;
     }
 
-    public void setPreActiveToLive(Float preActiveToLive) {
+    public void setPreActiveToLive(Double preActiveToLive) {
         this.preActiveToLive = preActiveToLive;
     }
 
-    public Float getAdmActiveToCanceled() {
+    public Double getAdmActiveToCanceled() {
         return admActiveToCanceled;
     }
 
-    public void setAdmActiveToCanceled(Float admActiveToCanceled) {
+    public void setAdmActiveToCanceled(Double admActiveToCanceled) {
         this.admActiveToCanceled = admActiveToCanceled;
     }
 
-    public Float getAdmActiveToSuspended() {
+    public Double getAdmActiveToSuspended() {
         return admActiveToSuspended;
     }
 
-    public void setAdmActiveToSuspended(Float admActiveToSuspended) {
+    public void setAdmActiveToSuspended(Double admActiveToSuspended) {
         this.admActiveToSuspended = admActiveToSuspended;
     }
 
-    public Float getAdmSuspendedToActive() {
+    public Double getAdmSuspendedToActive() {
         return admSuspendedToActive;
     }
 
-    public void setAdmSuspendedToActive(Float admSuspendedToActive) {
+    public void setAdmSuspendedToActive(Double admSuspendedToActive) {
         this.admSuspendedToActive = admSuspendedToActive;
     }
 
-    public Float getAdmSuspendedToCanceled() {
+    public Double getAdmSuspendedToCanceled() {
         return admSuspendedToCanceled;
     }
 
-    public void setAdmSuspendedToCanceled(Float admSuspendedToCanceled) {
+    public void setAdmSuspendedToCanceled(Double admSuspendedToCanceled) {
         this.admSuspendedToCanceled = admSuspendedToCanceled;
     }
